PIMS-322 update landing page.
