+++ frontend/Jenkinsfile.pipeline.param
@@ -7,6 +7,6 @@ NAME=frontend
 # GITHUB_WEBHOOK_SECRET=[a-zA-Z0-9]{40}
 # GENERIC_WEBHOOK_SECRET=[a-zA-Z0-9]{40}
 SOURCE_REPOSITORY_URL=https://github.com/devinleighsmith/PIMS.git
-SOURCE_REPOSITORY_REF=master
+SOURCE_REPOSITORY_REF=dev
 CONTEXT_DIR=./frontend
 JENKINSFILE_PATH=Jenkinsfile
