+++ frontend/openshift/frontend-build.param
@@ -8,7 +8,7 @@ ENV_NAME=jcxjin-tools
 BUILDER_IMAGE=yarn-builder
 RUNTIME_IMAGE=nginx-runtime
 GIT_REPO_URL=https://github.com/devinleighsmith/PIMS.git
-GIT_REF=master
+GIT_REF=dev
 SOURCE_CONTEXT_DIR=frontend
 JENKINS_FILEPATH=Jenkinsfile
 OUTPUT_IMAGE_TAG=latest
