+++ frontend/openshift/frontend-deploy.test.param
@@ -11,5 +11,5 @@
 # API_URL=
 # API_SERVICE_NAME=api
 # API_PATH=/api
-APPLICATION_DOMAIN=frontend-jcxjin-test.pathfinder.gov.bc.ca
+APPLICATION_DOMAIN=pims-test.pathfinder.gov.bc.ca
 TAG_NAME=test
