+++ frontend/openshift/nginx-runtime.param
@@ -5,6 +5,6 @@
 #=========================================================
 NAME=nginx-runtime
 GIT_REPO_URL=https://github.com/devinleighsmith/PIMS.git
-GIT_REF=master
+GIT_REF=dev
 SOURCE_CONTEXT_DIR=frontend/openshift/templates/nginx-runtime/
 OUTPUT_IMAGE_TAG=latest
