+++ frontend/openshift/templates/frontend/frontend-build.yaml
@@ -159,7 +159,7 @@ parameters:
     displayName: Git Reference
     description: The git reference or branch.
     required: true
-    value: master
+    value: dev
   - name: SOURCE_CONTEXT_DIR
     displayName: Source Context Directory
     description: The source context directory.
