+++ frontend/openshift/templates/yarn-builder/yarn-builder-build.json
@@ -70,7 +70,7 @@
       "displayName": "Git Reference",
       "description": "The git reference or branch.",
       "required": true,
-      "value": "master"
+      "value": "dev"
     },
     {
       "name": "SOURCE_CONTEXT_DIR",
