+++ frontend/openshift/yarn-builder-build.param
@@ -5,6 +5,6 @@
 #=========================================================
 NAME=yarn-builder
 GIT_REPO_URL=https://github.com/devinleighsmith/PIMS.git
-GIT_REF=master
+GIT_REF=dev
 SOURCE_CONTEXT_DIR=frontend/openshift/templates/yarn-builder/
 OUTPUT_IMAGE_TAG=latest
