+++ frontend/package.json
@@ -28,7 +28,7 @@
     "react-leaflet": "~2.6.0",
     "react-router-dom": "^5.1.2",
     "react-scripts": "3.3.0",
-    "typescript": "~3.7.3"
+    "typescript": "3.7.3"
   },
   "devDependencies": {
     "eslint-config-prettier": "~6.7.0",
