+++ frontend/public/keycloak.json
@@ -1,8 +1,8 @@
 {
-  "realm": "quartech",
-  "auth-server-url": "http://localhost:8080/auth",
+  "realm": "xz0xtue5",
+  "auth-server-url": "https://sso-dev.pathfinder.gov.bc.ca/auth",
   "ssl-required": "external",
-  "resource": "team-blue-app",
+  "resource": "pims-app",
   "public-client": true,
   "confidential-port": 0
 }
