--- frontend/src/components/maps/EditPlaceForm.scss
@@ -1,68 +1,0 @@
-.modal-body {
-    display: flex;
-  p {
-      widows: 324px;
-      height: 37px;
-  }
-  img {
-      margin-right: 21px;
-  }
-}
-.edit-place {
-  width: 212px;
-  height: 180px;
-  margin: 25px;
-
-  form {
-    font-family: ‘Noto Sans’, Verdana, Arial, sans-serif;
-    font-size: 18px;
-  }
-
-  h1 {
-    text-align: left;
-    height: 19px;
-    margin: 0px;
-    font-family: Arial;
-    font-size: 16px;
-    color: #606060;
-    text-decoration: none solid rgb(96, 96, 96);
-  }
-
-  .button {
-    display: block;
-    width: 82px;
-    box-sizing: border-box;
-    margin: 0;
-    padding: 10px 16px;
-    width: 100%;
-    outline: 0;
-    border: none;
-    border-color: none;
-    border-radius: 0;
-    color: #fff;
-    font-family: Arial;
-    text-align: center;
-    text-decoration: none;
-    font-size: 16px;
-    line-height: 1.3333333;
-    cursor: pointer;
-    border-radius: 50% !important;
-  }
-
-  .BC-Gov-PrimaryButton {
-    background-color: #38598a;
-    border-radius: 4px;
-    color: white;
-    display: inline-block;
-    margin-left: 48px;
-  }
-
-  .button:hover {
-    background: rgba(54, 50, 119, 0.8);
-  }
-
-  .button:focus {
-    border-color: #0074c2;
-    box-shadow: 0 0 8px rgba(0, 116, 194, 0.6);
-  }
-}
