--- frontend/src/components/navigation/Footer.js
@@ -1,14 +1,0 @@
-import './Footer.scss';
-import React from 'react';
-
-function Footer(props) {
-
-  return (
-    <footer className="footer">
-    </footer>
-  );
-}
-
-export default Footer;
-
-
