+++ frontend/src/components/navigation/Footer.scss
@@ -3,7 +3,25 @@
   color: #fff;
   bottom: 0;
   width: 100%;
-  padding: 20px 15px;
-  text-align: center;
-  position: absolute;
+  padding: 10px 15px;
+  border-top: 2px solid #fcba19;
+  .content {
+    margin: 0 80% 0 20%;
+  }
+
+  a:link{
+    color: #ffffff;
+    font-size: 16px;
+    font-family: "BCSans" Fallback, sans-serif;
+    padding: 0 10px;
+    border-right: 1px solid #666;
+  }
+
+  a:visited{
+    color: #ffffff;
+    font-size: 16px;
+    font-family: "BCSans" Fallback, sans-serif;
+    padding: 0 10px;
+    border-right: 1px solid #666;
+  }
 }
