+++ frontend/src/components/navigation/Footer.tsx
@@ -1,0 +1,20 @@
+import './Footer.scss';
+import React from 'react';
+
+function Footer() {
+
+  return (
+    <footer className="footer">
+      <div className="content">
+        <a href="http://www.gov.bc.ca/gov/content/home/disclaimer">Disclaimer</a>
+        <a href="http://www.gov.bc.ca/gov/content/home/privacy">Privacy</a>
+        <a href="http://www.gov.bc.ca/gov/content/home/accessible-government">Accessibility</a>
+        <a href="http://www.gov.bc.ca/gov/content/home/copyright">Copyright</a>
+      </div>
+    </footer>
+  );
+}
+
+export default Footer;
+
+
