+++ frontend/src/fonts.scss
@@ -1,0 +1,21 @@
+@font-face {
+  font-family: "BCSans";
+  font-style: normal;
+  src: url("../node_modules/@bcgov/bc-sans/fonts/BCSans-Regular.woff") format("woff");
+}
+@font-face {
+  font-family: "BCSans";
+  font-style: italic;
+  src: url("../node_modules/@bcgov/bc-sans/fonts/BCSans-Italic.woff") format("woff");
+}
+@font-face {
+  font-family: "BCSans";
+  font-weight: 700;
+  src: url("../node_modules/@bcgov/bc-sans/fonts/BCSans-Bold.woff") format("woff");
+}
+@font-face {
+  font-family: "BCSans";
+  font-style: italic;
+  font-weight: 700;
+  src: url("../node_modules/@bcgov/bc-sans/fonts/BCSans-BoldItalic.woff") format("woff");
+}
