+++ frontend/src/index.tsx
@@ -3,12 +3,12 @@ import ReactDOM from 'react-dom';
 import './index.scss';
 import App from './App';
 import * as serviceWorker from './serviceWorker';
-import Keycloak from 'keycloak-js';
+import Keycloak, { KeycloakInstance } from 'keycloak-js';
 import { KeycloakProvider } from 'react-keycloak';
 import { Spinner } from 'react-bootstrap';
-import { KEYCLOAK } from 'constants/environment'
 
-const keycloak = new Keycloak(KEYCLOAK);
+//@ts-ignore
+const keycloak:KeycloakInstance = new Keycloak('keycloak.json');
 const Index = () => {
     return <KeycloakProvider
         keycloak={keycloak}
