+++ frontend/src/pages/Login.scss
@@ -1,33 +1,44 @@
-.unauth {
-  width: 752px;
-  height: 254px;
-  margin: 180px auto;
-  background-color: #38598a;
-  padding: 30px 20px;
+@import "../fonts.scss";
+@import "../buttons.scss";
 
-  h1 {
-    margin-top: 30px;
-    padding-bottom: 35px;
-    font-family: Arial;
-    font-size: 31px;
-    color: #ffffff;
-    text-decoration: none solid rgb(255, 255, 255);
+.unauth {
+  width: 776px;
+  font-size: 24px;
+  margin: 91px auto;
+
+  .block {
+    width: 776px;
+    height: 332px;
+    background-color: #f2f2f2;
+    padding-top: 43px;
   }
 
-  .button {
-    font-family: Arial;
+  h1 {
+    padding-bottom: 18px;
+    font-family: 'BCSans' Fallback, sans-serif;
+    font-weight: 700;
     font-size: 24px;
     color: #494949;
-    text-decoration: none solid rgb(73, 73, 73);
-    background-color: white;
-    border: none;
-    border-radius: 4px;
-    padding: 8px 24px;
-    text-align: center;
-    display: inline-block;
-    font-size: 24px;
-    font-family: Arial;
-    letter-spacing: 1px;
-    cursor: pointer;
+  }
+
+  p {
+    font-size: 16px;
+    color: #494949;
+    font-family: 'BCSans' Fallback, sans-serif;
+    line-height: 130%;
+  }
+
+  .blockText {
+    margin: 0 auto 18px auto;
+    width: 568px;
+    height: 80px;
+  }
+  
+  .or {
+    margin: 20px auto;
+  }
+
+  .border-dark {
+    border-color: #003366 !important;
   }
 }
